test (core module): test issues (not fixed yet)
