+++ projects/knora/core/src/lib/test-data/admin/shared-test-data.ts
@@ -51,7 +51,7 @@ export const imagesReviewerGroup: Group = imagesReviewerGroupResponse.group;
 
 
 // ---- Users ----
-export const usersResponseJson: any = { 'users': [{ 'familyName': 'Admin-alt', 'givenName': 'Administrator-alt', 'email': 'root-alt@example.com', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/91e19f1e01', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'User', 'givenName': 'User02', 'email': 'user02.user@example.com', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/97cec4000f', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'User01', 'givenName': 'Anything', 'email': 'anything.user01@example.org', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/9XBCrDV3SRa7kS1WwynB4Q', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'Admin', 'givenName': 'Anything', 'email': 'anything.admin@example.org', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/AnythingAdminUser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'User02', 'givenName': 'Anything', 'email': 'anything.user02@example.org', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/BhkfBc3hTeS_IDo-JgXRbQ', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'BEOL', 'givenName': 'BEOL', 'email': 'beol@example.com', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/PSGbemdjZi4kQ6GHJVkLGE', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'BEOL', 'givenName': 'BEOL', 'email': 't.schweizer@unibas.ch', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/PSGbemdjZi4kQ6GHJVkLGF', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'biblio', 'givenName': 'biblio', 'email': 'biblio@example.com', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/Q-6Sssu8TBWrcCGuVJ0lVw', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'Test', 'givenName': 'User', 'email': 'user.test@example.com', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/b83acc5f05', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'User', 'givenName': 'User01', 'email': 'user01.user1@example.com', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/c266a56709', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'User', 'givenName': 'User03', 'email': 'images-reviewer-user@example.com', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/images-reviewer-user', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'User', 'givenName': 'Inactive', 'email': 'inactive.user@example.com', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/inactiveuser', 'status': false, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'User2', 'givenName': 'Test', 'email': 'test.user2@test.ch', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/incunabulaMemberUser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'User', 'givenName': 'Multi', 'email': 'multi.user@example.com', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/multiuser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'User', 'givenName': 'Normal', 'email': 'normal.user@example.com', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/normaluser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'Administrator', 'givenName': 'System', 'email': 'root@example.com', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/root', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'User', 'givenName': 'Super', 'email': 'super.user@example.com', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/superuser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'Webern', 'givenName': 'Admin', 'email': 'webern-admin@example.ch', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/webernProjectAdmin', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'Webern', 'givenName': 'Nutzer', 'email': 'webern-nutzer@example.ch', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://rdfh.ch/users/webernProjectMember', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null }, { 'familyName': 'Anonymous', 'givenName': 'Knora', 'email': 'anonymous@localhost', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://www.knora.org/ontology/knora-base#AnonymousUser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'en', 'password': null }, { 'familyName': 'System', 'givenName': 'Knora', 'email': 'system@localhost', 'permissions': { 'groupsPerProject': {}, 'administrativePermissionsPerProject': {} }, 'groups': [], 'id': 'http://www.knora.org/ontology/knora-base#SystemUser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'en', 'password': null }] };
+export const usersResponseJson: any = {'users': [{'familyName': 'Admin-alt', 'givenName': 'Administrator-alt', 'email': 'root-alt@example.com', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/91e19f1e01', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'User', 'givenName': 'User02', 'email': 'user02.user@example.com', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/97cec4000f', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'User01', 'givenName': 'Anything', 'email': 'anything.user01@example.org', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/9XBCrDV3SRa7kS1WwynB4Q', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'Admin', 'givenName': 'Anything', 'email': 'anything.admin@example.org', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/AnythingAdminUser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'User02', 'givenName': 'Anything', 'email': 'anything.user02@example.org', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/BhkfBc3hTeS_IDo-JgXRbQ', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'BEOL', 'givenName': 'BEOL', 'email': 'beol@example.com', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/PSGbemdjZi4kQ6GHJVkLGE', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'BEOL', 'givenName': 'BEOL', 'email': 't.schweizer@unibas.ch', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/PSGbemdjZi4kQ6GHJVkLGF', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'biblio', 'givenName': 'biblio', 'email': 'biblio@example.com', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/Q-6Sssu8TBWrcCGuVJ0lVw', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'Test', 'givenName': 'User', 'email': 'user.test@example.com', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/b83acc5f05', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'User', 'givenName': 'User01', 'email': 'user01.user1@example.com', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/c266a56709', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'User', 'givenName': 'User03', 'email': 'images-reviewer-user@example.com', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/images-reviewer-user', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'User', 'givenName': 'Inactive', 'email': 'inactive.user@example.com', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/inactiveuser', 'status': false, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'User2', 'givenName': 'Test', 'email': 'test.user2@test.ch', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/incunabulaMemberUser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'User', 'givenName': 'Multi', 'email': 'multi.user@example.com', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/multiuser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'User', 'givenName': 'Normal', 'email': 'normal.user@example.com', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/normaluser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'Administrator', 'givenName': 'System', 'email': 'root@example.com', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/root', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'User', 'givenName': 'Super', 'email': 'super.user@example.com', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/superuser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'Webern', 'givenName': 'Admin', 'email': 'webern-admin@example.ch', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/webernProjectAdmin', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'Webern', 'givenName': 'Nutzer', 'email': 'webern-nutzer@example.ch', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://rdfh.ch/users/webernProjectMember', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'de', 'password': null}, {'familyName': 'Anonymous', 'givenName': 'Knora', 'email': 'anonymous@localhost', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://www.knora.org/ontology/knora-base#AnonymousUser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'en', 'password': null}, {'familyName': 'System', 'givenName': 'Knora', 'email': 'system@localhost', 'permissions': {'groupsPerProject': {}, 'administrativePermissionsPerProject': {}}, 'groups': [], 'id': 'http://www.knora.org/ontology/knora-base#SystemUser', 'status': true, 'token': null, 'sessionId': null, 'projects': [], 'lang': 'en', 'password': null}]};
 export const usersResponse: UsersResponse = jsonConvert.deserializeObject(usersResponseJson, UsersResponse);
 export const usersTestData: User[] = usersResponse.users;
